Update testq.go
Added functions for deleting internal and global orders.
Was thinking of making a function to delete orders in my_order_list, but it'll make chaos. When deleting an order I just shift the orders to the left over the order to be deleted. If we do this in the big list we risk mixing up the two lists (getting orders from global to internal etc.). It's easier just deleting in the correct list then make my_order_list again with the updated lists.